+++ components/MortgageCalculator/InterestRateInput.module.css
@@ -1,0 +1,11 @@
+.sliderValue {
+  display: flex;
+  flex-direction: row;
+  align-items: flex-start;
+  font-size: 1.8rem;
+  margin: 0.4rem 0;
+}
+
+.sliderValue.interestRateSliderValue {
+  margin-left: 1rem;
+}
