+++ components/MortgageCalculator/MortgageCalculatorCard.module.css
@@ -1,67 +1,3 @@
-.container {
-  background-color: var(--background-color);
-}
-
-.content {
-  padding: 50px 0 100px;
-}
-
-.heading {
-  color: var(--color-heading);
-  margin-bottom: 0;
-  font-size: 1.7rem;
-  font-weight: 400;
-}
-
-.subtext {
-  color: var(--color-tertiary);
-  margin-bottom: 3rem;
-  margin-top: 0.5rem;
-  font-size: 0.9rem;
-}
-
-.grid {
-  display: grid;
-  grid-template-columns: 1fr;
-  grid-gap: 2rem;
-}
-
-@media (min-width: 768px) {
-  .grid {
-    grid-template-columns: 1fr 1fr;
-    grid-gap: 0;
-  }
-
-  .gridLeft {
-    padding-right: 20%;
-  }
-}
-
-.gridRight {
-  position: relative;
-}
-
-.sliderValue {
-  display: flex;
-  flex-direction: row;
-  align-items: flex-start;
-  font-size: 1.8rem;
-  margin: 0.4rem 0;
-}
-
-.sliderValue .sliderValueCurrency {
-  font-size: 1rem;
-  margin-right: 0.2rem;
-}
-
-.sliderValue .sliderValueNumber {
-  line-height: 2rem;
-}
-
-.sliderValue.interestRateSliderValue {
-  margin-left: 1rem;
-}
-
 .card {
   border-radius: var(--border-radius);
   display: flex;
