+++ pages/index.tsx
@@ -1,5 +1,5 @@
 import type { NextPage } from "next";
-import MortgageCalculator from "../components/MortgageCalculator";
+import MortgageCalculator from "../components/MortgageCalculator/MortgageCalculator";
 
 const Home: NextPage = () => {
   return (
